requirements.txt
# Python dependencies
gradio>=4.0.0
smolagents>=0.1.0
fal-client>=0.4.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0
